refactoring:new Todos component and changed todos functions names
